Calculate ELSA scores

Calculating the Entropy-based Local indicator of Spatial Association (ELSA) scores, which consist of Ea, Ec and the overall ELSA.

Usage

ELSA(labels, location, k = 10)

Arguments

labels

Cluster labels.

location

A numerical matrix containing the location information, with rows as samples and columns as location dimensions.

k

Number of nearest neighbors.

Value

A dataframe containing the Ea, Ec and ELSA for all samples in the dataset.

References

Naimi, Babak, et al., 2019; 10.1016/j.spasta.2018.10.001

Examples

data(sp_toys)
data <- sp_toys
ELSA(data$label, data[,c("x", "y")], k=6)
#> the specified variable is considered as categorical...
#>             Ea        Ec       ELSA
#> 1   0.00000000 0.0000000 0.00000000
#> 2   0.00000000 0.0000000 0.00000000
#> 3   0.09090909 0.4138169 0.03761971
#> 4   0.20000000 0.6840384 0.13680769
#> 5   0.66666667 0.9709506 0.64730040
#> 6   0.11111111 0.4689956 0.05211062
#> 7   0.00000000 0.0000000 0.00000000
#> 8   0.00000000 0.0000000 0.00000000
#> 9   0.00000000 0.0000000 0.00000000
#> 10  0.00000000 0.0000000 0.00000000
#> 11  0.00000000 0.0000000 0.00000000
#> 12  0.00000000 0.0000000 0.00000000
#> 13  0.00000000 0.0000000 0.00000000
#> 14  0.00000000 0.0000000 0.00000000
#> 15  0.00000000 0.0000000 0.00000000
#> 16  0.00000000 0.0000000 0.00000000
#> 17  0.00000000 0.0000000 0.00000000
#> 18  0.13333333 0.5435644 0.07247526
#> 19  0.28571429 0.8366407 0.23904021
#> 20  0.53846154 1.0000000 0.53846154
#> 21  0.07692308 0.3712323 0.02855633
#> 22  0.00000000 0.0000000 0.00000000
#> 23  0.00000000 0.0000000 0.00000000
#> 24  0.00000000 0.0000000 0.00000000
#> 25  0.00000000 0.0000000 0.00000000
#> 26  0.00000000 0.0000000 0.00000000
#> 27  0.00000000 0.0000000 0.00000000
#> 28  0.00000000 0.0000000 0.00000000
#> 29  0.00000000 0.0000000 0.00000000
#> 30  0.00000000 0.0000000 0.00000000
#> 31  0.00000000 0.0000000 0.00000000
#> 32  0.00000000 0.0000000 0.00000000
#> 33  0.00000000 0.0000000 0.00000000
#> 34  0.11764706 0.5032583 0.05920686
#> 35  0.33333333 0.8960382 0.29867941
#> 36  0.33333333 0.8960382 0.29867941
#> 37  0.11764706 0.5032583 0.05920686
#> 38  0.00000000 0.0000000 0.00000000
#> 39  0.00000000 0.0000000 0.00000000
#> 40  0.00000000 0.0000000 0.00000000
#> 41  0.00000000 0.0000000 0.00000000
#> 42  0.00000000 0.0000000 0.00000000
#> 43  0.00000000 0.0000000 0.00000000
#> 44  0.00000000 0.0000000 0.00000000
#> 45  0.00000000 0.0000000 0.00000000
#> 46  0.00000000 0.0000000 0.00000000
#> 47  0.00000000 0.0000000 0.00000000
#> 48  0.00000000 0.0000000 0.00000000
#> 49  0.11764706 0.5032583 0.05920686
#> 50  0.33333333 0.8960382 0.29867941
#> 51  0.33333333 0.8960382 0.29867941
#> 52  0.11764706 0.5032583 0.05920686
#> 53  0.00000000 0.0000000 0.00000000
#> 54  0.00000000 0.0000000 0.00000000
#> 55  0.00000000 0.0000000 0.00000000
#> 56  0.00000000 0.0000000 0.00000000
#> 57  0.00000000 0.0000000 0.00000000
#> 58  0.00000000 0.0000000 0.00000000
#> 59  0.00000000 0.0000000 0.00000000
#> 60  0.00000000 0.0000000 0.00000000
#> 61  0.00000000 0.0000000 0.00000000
#> 62  0.00000000 0.0000000 0.00000000
#> 63  0.00000000 0.0000000 0.00000000
#> 64  0.00000000 0.0000000 0.00000000
#> 65  0.06666667 0.3372901 0.02248600
#> 66  0.46666667 0.9886994 0.46139306
#> 67  0.35294118 0.9182958 0.32410441
#> 68  0.16666667 0.6292492 0.10487487
#> 69  0.00000000 0.0000000 0.00000000
#> 70  0.00000000 0.0000000 0.00000000
#> 71  0.00000000 0.0000000 0.00000000
#> 72  0.00000000 0.0000000 0.00000000
#> 73  0.00000000 0.0000000 0.00000000
#> 74  0.00000000 0.0000000 0.00000000
#> 75  0.00000000 0.0000000 0.00000000
#> 76  0.00000000 0.0000000 0.00000000
#> 77  0.00000000 0.0000000 0.00000000
#> 78  0.00000000 0.0000000 0.00000000
#> 79  0.00000000 0.0000000 0.00000000
#> 80  0.06666667 0.3372901 0.02248600
#> 81  0.46666667 0.9886994 0.46139306
#> 82  0.35294118 0.9182958 0.32410441
#> 83  0.16666667 0.6292492 0.10487487
#> 84  0.00000000 0.0000000 0.00000000
#> 85  0.00000000 0.0000000 0.00000000
#> 86  0.00000000 0.0000000 0.00000000
#> 87  0.00000000 0.0000000 0.00000000
#> 88  0.00000000 0.0000000 0.00000000
#> 89  0.00000000 0.0000000 0.00000000
#> 90  0.00000000 0.0000000 0.00000000
#> 91  0.00000000 0.0000000 0.00000000
#> 92  0.00000000 0.0000000 0.00000000
#> 93  0.00000000 0.0000000 0.00000000
#> 94  0.00000000 0.0000000 0.00000000
#> 95  0.00000000 0.0000000 0.00000000
#> 96  0.20000000 0.6962123 0.13924245
#> 97  0.41176471 0.9640788 0.39697361
#> 98  0.38888889 0.9494520 0.36923134
#> 99  0.17647059 0.6500224 0.11470984
#> 100 0.00000000 0.0000000 0.00000000
#> 101 0.00000000 0.0000000 0.00000000
#> 102 0.00000000 0.0000000 0.00000000
#> 103 0.00000000 0.0000000 0.00000000
#> 104 0.00000000 0.0000000 0.00000000
#> 105 0.00000000 0.0000000 0.00000000
#> 106 0.00000000 0.0000000 0.00000000
#> 107 0.00000000 0.0000000 0.00000000
#> 108 0.00000000 0.0000000 0.00000000
#> 109 0.00000000 0.0000000 0.00000000
#> 110 0.00000000 0.0000000 0.00000000
#> 111 0.20000000 0.6962123 0.13924245
#> 112 0.41176471 0.9640788 0.39697361
#> 113 0.38888889 0.9494520 0.36923134
#> 114 0.16666667 0.6292492 0.10487487
#> 115 0.00000000 0.0000000 0.00000000
#> 116 0.00000000 0.0000000 0.00000000
#> 117 0.00000000 0.0000000 0.00000000
#> 118 0.00000000 0.0000000 0.00000000
#> 119 0.00000000 0.0000000 0.00000000
#> 120 0.00000000 0.0000000 0.00000000
#> 121 0.00000000 0.0000000 0.00000000
#> 122 0.00000000 0.0000000 0.00000000
#> 123 0.00000000 0.0000000 0.00000000
#> 124 0.00000000 0.0000000 0.00000000
#> 125 0.00000000 0.0000000 0.00000000
#> 126 0.00000000 0.0000000 0.00000000
#> 127 0.17647059 0.6500224 0.11470984
#> 128 0.38888889 0.9494520 0.36923134
#> 129 0.41176471 0.9640788 0.39697361
#> 130 0.20000000 0.6962123 0.13924245
#> 131 0.00000000 0.0000000 0.00000000
#> 132 0.00000000 0.0000000 0.00000000
#> 133 0.00000000 0.0000000 0.00000000
#> 134 0.00000000 0.0000000 0.00000000
#> 135 0.00000000 0.0000000 0.00000000
#> 136 0.00000000 0.0000000 0.00000000
#> 137 0.00000000 0.0000000 0.00000000
#> 138 0.00000000 0.0000000 0.00000000
#> 139 0.00000000 0.0000000 0.00000000
#> 140 0.00000000 0.0000000 0.00000000
#> 141 0.00000000 0.0000000 0.00000000
#> 142 0.17647059 0.6500224 0.11470984
#> 143 0.38888889 0.9494520 0.36923134
#> 144 0.38888889 0.9494520 0.36923134
#> 145 0.20000000 0.6962123 0.13924245
#> 146 0.00000000 0.0000000 0.00000000
#> 147 0.00000000 0.0000000 0.00000000
#> 148 0.00000000 0.0000000 0.00000000
#> 149 0.00000000 0.0000000 0.00000000
#> 150 0.00000000 0.0000000 0.00000000
#> 151 0.00000000 0.0000000 0.00000000
#> 152 0.00000000 0.0000000 0.00000000
#> 153 0.00000000 0.0000000 0.00000000
#> 154 0.00000000 0.0000000 0.00000000
#> 155 0.00000000 0.0000000 0.00000000
#> 156 0.00000000 0.0000000 0.00000000
#> 157 0.00000000 0.0000000 0.00000000
#> 158 0.16666667 0.6292492 0.10487487
#> 159 0.35294118 0.9182958 0.32410441
#> 160 0.46666667 0.9886994 0.46139306
#> 161 0.07692308 0.3712323 0.02855633
#> 162 0.00000000 0.0000000 0.00000000
#> 163 0.00000000 0.0000000 0.00000000
#> 164 0.00000000 0.0000000 0.00000000
#> 165 0.00000000 0.0000000 0.00000000
#> 166 0.00000000 0.0000000 0.00000000
#> 167 0.00000000 0.0000000 0.00000000
#> 168 0.00000000 0.0000000 0.00000000
#> 169 0.00000000 0.0000000 0.00000000
#> 170 0.00000000 0.0000000 0.00000000
#> 171 0.00000000 0.0000000 0.00000000
#> 172 0.00000000 0.0000000 0.00000000
#> 173 0.16666667 0.6292492 0.10487487
#> 174 0.38888889 0.9494520 0.36923134
#> 175 0.46666667 0.9886994 0.46139306
#> 176 0.14285714 0.5665095 0.08092993
#> 177 0.00000000 0.0000000 0.00000000
#> 178 0.00000000 0.0000000 0.00000000
#> 179 0.00000000 0.0000000 0.00000000
#> 180 0.00000000 0.0000000 0.00000000
#> 181 0.00000000 0.0000000 0.00000000
#> 182 0.00000000 0.0000000 0.00000000
#> 183 0.00000000 0.0000000 0.00000000
#> 184 0.00000000 0.0000000 0.00000000
#> 185 0.00000000 0.0000000 0.00000000
#> 186 0.00000000 0.0000000 0.00000000
#> 187 0.00000000 0.0000000 0.00000000
#> 188 0.00000000 0.0000000 0.00000000
#> 189 0.11764706 0.5032583 0.05920686
#> 190 0.33333333 0.8960382 0.29867941
#> 191 0.38461538 0.9402860 0.36164845
#> 192 0.06666667 0.3372901 0.02248600
#> 193 0.00000000 0.0000000 0.00000000
#> 194 0.00000000 0.0000000 0.00000000
#> 195 0.00000000 0.0000000 0.00000000
#> 196 0.00000000 0.0000000 0.00000000
#> 197 0.00000000 0.0000000 0.00000000
#> 198 0.00000000 0.0000000 0.00000000
#> 199 0.00000000 0.0000000 0.00000000
#> 200 0.00000000 0.0000000 0.00000000
#> 201 0.00000000 0.0000000 0.00000000
#> 202 0.00000000 0.0000000 0.00000000
#> 203 0.00000000 0.0000000 0.00000000
#> 204 0.16666667 0.6292492 0.10487487
#> 205 0.33333333 0.8960382 0.29867941
#> 206 0.42857143 0.9709506 0.41612168
#> 207 0.06666667 0.3372901 0.02248600
#> 208 0.00000000 0.0000000 0.00000000
#> 209 0.00000000 0.0000000 0.00000000
#> 210 0.00000000 0.0000000 0.00000000
#> 211 0.00000000 0.0000000 0.00000000
#> 212 0.00000000 0.0000000 0.00000000
#> 213 0.00000000 0.0000000 0.00000000
#> 214 0.00000000 0.0000000 0.00000000
#> 215 0.00000000 0.0000000 0.00000000
#> 216 0.00000000 0.0000000 0.00000000
#> 217 0.00000000 0.0000000 0.00000000
#> 218 0.00000000 0.0000000 0.00000000
#> 219 0.00000000 0.0000000 0.00000000
#> 220 0.07692308 0.3712323 0.02855633
#> 221 0.50000000 0.9957275 0.49786373
#> 222 0.30769231 0.8631206 0.26557556
#> 223 0.14285714 0.5665095 0.08092993
#> 224 0.00000000 0.0000000 0.00000000
#> 225 0.00000000 0.0000000 0.00000000
#> 226 0.00000000 0.0000000 0.00000000
#> 227 0.00000000 0.0000000 0.00000000
#> 228 0.00000000 0.0000000 0.00000000
#> 229 0.00000000 0.0000000 0.00000000
#> 230 0.00000000 0.0000000 0.00000000
#> 231 0.00000000 0.0000000 0.00000000
#> 232 0.00000000 0.0000000 0.00000000
#> 233 0.00000000 0.0000000 0.00000000
#> 234 0.00000000 0.0000000 0.00000000
#> 235 0.11111111 0.4689956 0.05211062
#> 236 0.55555556 1.0000000 0.55555556
#> 237 0.22222222 0.7219281 0.16042847
#> 238 0.10000000 0.4394970 0.04394970
#> 239 0.00000000 0.0000000 0.00000000
#> 240 0.00000000 0.0000000 0.00000000
ELSA(data$p1, data[,c("x", "y")], k=6)
#> the specified variable is considered as categorical...
#>             Ea        Ec       ELSA
#> 1   0.00000000 0.0000000 0.00000000
#> 2   0.00000000 0.0000000 0.00000000
#> 3   0.00000000 0.0000000 0.00000000
#> 4   0.00000000 0.0000000 0.00000000
#> 5   0.22222222 0.7219281 0.16042847
#> 6   0.44444444 0.9709506 0.43153360
#> 7   0.20000000 0.6840384 0.13680769
#> 8   0.00000000 0.0000000 0.00000000
#> 9   0.00000000 0.0000000 0.00000000
#> 10  0.00000000 0.0000000 0.00000000
#> 11  0.00000000 0.0000000 0.00000000
#> 12  0.00000000 0.0000000 0.00000000
#> 13  0.00000000 0.0000000 0.00000000
#> 14  0.00000000 0.0000000 0.00000000
#> 15  0.00000000 0.0000000 0.00000000
#> 16  0.00000000 0.0000000 0.00000000
#> 17  0.00000000 0.0000000 0.00000000
#> 18  0.00000000 0.0000000 0.00000000
#> 19  0.07142857 0.3533593 0.02523995
#> 20  0.30769231 0.8631206 0.26557556
#> 21  0.38461538 0.9402860 0.36164845
#> 22  0.14285714 0.5665095 0.08092993
#> 23  0.00000000 0.0000000 0.00000000
#> 24  0.00000000 0.0000000 0.00000000
#> 25  0.00000000 0.0000000 0.00000000
#> 26  0.00000000 0.0000000 0.00000000
#> 27  0.00000000 0.0000000 0.00000000
#> 28  0.00000000 0.0000000 0.00000000
#> 29  0.00000000 0.0000000 0.00000000
#> 30  0.00000000 0.0000000 0.00000000
#> 31  0.00000000 0.0000000 0.00000000
#> 32  0.00000000 0.0000000 0.00000000
#> 33  0.00000000 0.0000000 0.00000000
#> 34  0.00000000 0.0000000 0.00000000
#> 35  0.06666667 0.3372901 0.02248600
#> 36  0.46666667 0.9886994 0.46139306
#> 37  0.35294118 0.9182958 0.32410441
#> 38  0.16666667 0.6292492 0.10487487
#> 39  0.00000000 0.0000000 0.00000000
#> 40  0.00000000 0.0000000 0.00000000
#> 41  0.00000000 0.0000000 0.00000000
#> 42  0.00000000 0.0000000 0.00000000
#> 43  0.00000000 0.0000000 0.00000000
#> 44  0.00000000 0.0000000 0.00000000
#> 45  0.00000000 0.0000000 0.00000000
#> 46  0.00000000 0.0000000 0.00000000
#> 47  0.00000000 0.0000000 0.00000000
#> 48  0.00000000 0.0000000 0.00000000
#> 49  0.00000000 0.0000000 0.00000000
#> 50  0.06666667 0.3372901 0.02248600
#> 51  0.46666667 0.9886994 0.46139306
#> 52  0.35294118 0.9182958 0.32410441
#> 53  0.16666667 0.6292492 0.10487487
#> 54  0.00000000 0.0000000 0.00000000
#> 55  0.00000000 0.0000000 0.00000000
#> 56  0.00000000 0.0000000 0.00000000
#> 57  0.00000000 0.0000000 0.00000000
#> 58  0.00000000 0.0000000 0.00000000
#> 59  0.00000000 0.0000000 0.00000000
#> 60  0.00000000 0.0000000 0.00000000
#> 61  0.00000000 0.0000000 0.00000000
#> 62  0.00000000 0.0000000 0.00000000
#> 63  0.00000000 0.0000000 0.00000000
#> 64  0.00000000 0.0000000 0.00000000
#> 65  0.00000000 0.0000000 0.00000000
#> 66  0.20000000 0.6962123 0.13924245
#> 67  0.41176471 0.9640788 0.39697361
#> 68  0.38888889 0.9494520 0.36923134
#> 69  0.17647059 0.6500224 0.11470984
#> 70  0.00000000 0.0000000 0.00000000
#> 71  0.00000000 0.0000000 0.00000000
#> 72  0.00000000 0.0000000 0.00000000
#> 73  0.00000000 0.0000000 0.00000000
#> 74  0.00000000 0.0000000 0.00000000
#> 75  0.00000000 0.0000000 0.00000000
#> 76  0.00000000 0.0000000 0.00000000
#> 77  0.00000000 0.0000000 0.00000000
#> 78  0.00000000 0.0000000 0.00000000
#> 79  0.00000000 0.0000000 0.00000000
#> 80  0.00000000 0.0000000 0.00000000
#> 81  0.20000000 0.6962123 0.13924245
#> 82  0.41176471 0.9640788 0.39697361
#> 83  0.38888889 0.9494520 0.36923134
#> 84  0.16666667 0.6292492 0.10487487
#> 85  0.00000000 0.0000000 0.00000000
#> 86  0.00000000 0.0000000 0.00000000
#> 87  0.00000000 0.0000000 0.00000000
#> 88  0.00000000 0.0000000 0.00000000
#> 89  0.00000000 0.0000000 0.00000000
#> 90  0.00000000 0.0000000 0.00000000
#> 91  0.00000000 0.0000000 0.00000000
#> 92  0.00000000 0.0000000 0.00000000
#> 93  0.00000000 0.0000000 0.00000000
#> 94  0.00000000 0.0000000 0.00000000
#> 95  0.00000000 0.0000000 0.00000000
#> 96  0.00000000 0.0000000 0.00000000
#> 97  0.17647059 0.6500224 0.11470984
#> 98  0.38888889 0.9494520 0.36923134
#> 99  0.41176471 0.9640788 0.39697361
#> 100 0.20000000 0.6962123 0.13924245
#> 101 0.00000000 0.0000000 0.00000000
#> 102 0.00000000 0.0000000 0.00000000
#> 103 0.00000000 0.0000000 0.00000000
#> 104 0.00000000 0.0000000 0.00000000
#> 105 0.00000000 0.0000000 0.00000000
#> 106 0.00000000 0.0000000 0.00000000
#> 107 0.00000000 0.0000000 0.00000000
#> 108 0.00000000 0.0000000 0.00000000
#> 109 0.00000000 0.0000000 0.00000000
#> 110 0.00000000 0.0000000 0.00000000
#> 111 0.00000000 0.0000000 0.00000000
#> 112 0.17647059 0.6500224 0.11470984
#> 113 0.38888889 0.9494520 0.36923134
#> 114 0.38888889 0.9494520 0.36923134
#> 115 0.20000000 0.6962123 0.13924245
#> 116 0.00000000 0.0000000 0.00000000
#> 117 0.00000000 0.0000000 0.00000000
#> 118 0.00000000 0.0000000 0.00000000
#> 119 0.00000000 0.0000000 0.00000000
#> 120 0.00000000 0.0000000 0.00000000
#> 121 0.00000000 0.0000000 0.00000000
#> 122 0.00000000 0.0000000 0.00000000
#> 123 0.00000000 0.0000000 0.00000000
#> 124 0.00000000 0.0000000 0.00000000
#> 125 0.00000000 0.0000000 0.00000000
#> 126 0.00000000 0.0000000 0.00000000
#> 127 0.00000000 0.0000000 0.00000000
#> 128 0.16666667 0.6292492 0.10487487
#> 129 0.35294118 0.9182958 0.32410441
#> 130 0.46666667 0.9886994 0.46139306
#> 131 0.07692308 0.3712323 0.02855633
#> 132 0.00000000 0.0000000 0.00000000
#> 133 0.00000000 0.0000000 0.00000000
#> 134 0.00000000 0.0000000 0.00000000
#> 135 0.00000000 0.0000000 0.00000000
#> 136 0.00000000 0.0000000 0.00000000
#> 137 0.00000000 0.0000000 0.00000000
#> 138 0.00000000 0.0000000 0.00000000
#> 139 0.00000000 0.0000000 0.00000000
#> 140 0.00000000 0.0000000 0.00000000
#> 141 0.00000000 0.0000000 0.00000000
#> 142 0.00000000 0.0000000 0.00000000
#> 143 0.16666667 0.6292492 0.10487487
#> 144 0.38888889 0.9494520 0.36923134
#> 145 0.46666667 0.9886994 0.46139306
#> 146 0.14285714 0.5665095 0.08092993
#> 147 0.00000000 0.0000000 0.00000000
#> 148 0.00000000 0.0000000 0.00000000
#> 149 0.00000000 0.0000000 0.00000000
#> 150 0.00000000 0.0000000 0.00000000
#> 151 0.00000000 0.0000000 0.00000000
#> 152 0.00000000 0.0000000 0.00000000
#> 153 0.00000000 0.0000000 0.00000000
#> 154 0.00000000 0.0000000 0.00000000
#> 155 0.00000000 0.0000000 0.00000000
#> 156 0.00000000 0.0000000 0.00000000
#> 157 0.00000000 0.0000000 0.00000000
#> 158 0.00000000 0.0000000 0.00000000
#> 159 0.11764706 0.5032583 0.05920686
#> 160 0.33333333 0.8960382 0.29867941
#> 161 0.38461538 0.9402860 0.36164845
#> 162 0.06666667 0.3372901 0.02248600
#> 163 0.00000000 0.0000000 0.00000000
#> 164 0.00000000 0.0000000 0.00000000
#> 165 0.00000000 0.0000000 0.00000000
#> 166 0.00000000 0.0000000 0.00000000
#> 167 0.00000000 0.0000000 0.00000000
#> 168 0.00000000 0.0000000 0.00000000
#> 169 0.00000000 0.0000000 0.00000000
#> 170 0.00000000 0.0000000 0.00000000
#> 171 0.00000000 0.0000000 0.00000000
#> 172 0.00000000 0.0000000 0.00000000
#> 173 0.00000000 0.0000000 0.00000000
#> 174 0.16666667 0.6292492 0.10487487
#> 175 0.33333333 0.8960382 0.29867941
#> 176 0.42857143 0.9709506 0.41612168
#> 177 0.06666667 0.3372901 0.02248600
#> 178 0.00000000 0.0000000 0.00000000
#> 179 0.00000000 0.0000000 0.00000000
#> 180 0.00000000 0.0000000 0.00000000
#> 181 0.00000000 0.0000000 0.00000000
#> 182 0.00000000 0.0000000 0.00000000
#> 183 0.00000000 0.0000000 0.00000000
#> 184 0.00000000 0.0000000 0.00000000
#> 185 0.00000000 0.0000000 0.00000000
#> 186 0.00000000 0.0000000 0.00000000
#> 187 0.00000000 0.0000000 0.00000000
#> 188 0.00000000 0.0000000 0.00000000
#> 189 0.00000000 0.0000000 0.00000000
#> 190 0.06666667 0.3372901 0.02248600
#> 191 0.46153846 0.9852281 0.45472068
#> 192 0.33333333 0.8960382 0.29867941
#> 193 0.18750000 0.6722948 0.12605528
#> 194 0.00000000 0.0000000 0.00000000
#> 195 0.00000000 0.0000000 0.00000000
#> 196 0.00000000 0.0000000 0.00000000
#> 197 0.00000000 0.0000000 0.00000000
#> 198 0.00000000 0.0000000 0.00000000
#> 199 0.00000000 0.0000000 0.00000000
#> 200 0.00000000 0.0000000 0.00000000
#> 201 0.00000000 0.0000000 0.00000000
#> 202 0.00000000 0.0000000 0.00000000
#> 203 0.00000000 0.0000000 0.00000000
#> 204 0.00000000 0.0000000 0.00000000
#> 205 0.06666667 0.3372901 0.02248600
#> 206 0.42857143 0.9709506 0.41612168
#> 207 0.33333333 0.8960382 0.29867941
#> 208 0.18750000 0.6722948 0.12605528
#> 209 0.00000000 0.0000000 0.00000000
#> 210 0.00000000 0.0000000 0.00000000
#> 211 0.00000000 0.0000000 0.00000000
#> 212 0.00000000 0.0000000 0.00000000
#> 213 0.00000000 0.0000000 0.00000000
#> 214 0.00000000 0.0000000 0.00000000
#> 215 0.00000000 0.0000000 0.00000000
#> 216 0.00000000 0.0000000 0.00000000
#> 217 0.00000000 0.0000000 0.00000000
#> 218 0.00000000 0.0000000 0.00000000
#> 219 0.00000000 0.0000000 0.00000000
#> 220 0.00000000 0.0000000 0.00000000
#> 221 0.16666667 0.6193822 0.10323037
#> 222 0.53846154 1.0000000 0.53846154
#> 223 0.35714286 0.9182958 0.32796280
#> 224 0.15384615 0.5916728 0.09102658
#> 225 0.00000000 0.0000000 0.00000000
#> 226 0.00000000 0.0000000 0.00000000
#> 227 0.00000000 0.0000000 0.00000000
#> 228 0.00000000 0.0000000 0.00000000
#> 229 0.00000000 0.0000000 0.00000000
#> 230 0.00000000 0.0000000 0.00000000
#> 231 0.00000000 0.0000000 0.00000000
#> 232 0.00000000 0.0000000 0.00000000
#> 233 0.00000000 0.0000000 0.00000000
#> 234 0.00000000 0.0000000 0.00000000
#> 235 0.00000000 0.0000000 0.00000000
#> 236 0.22222222 0.7219281 0.16042847
#> 237 0.66666667 0.9709506 0.64730040
#> 238 0.30000000 0.8453509 0.25360528
#> 239 0.12500000 0.5032583 0.06290729
#> 240 0.00000000 0.0000000 0.00000000
ELSA(data$p2, data[,c("x", "y")], k=6)
#> the specified variable is considered as categorical...
#>             Ea        Ec       ELSA
#> 1   0.00000000 0.0000000 0.00000000
#> 2   0.00000000 0.0000000 0.00000000
#> 3   0.09090909 0.4138169 0.03761971
#> 4   0.20000000 0.6840384 0.13680769
#> 5   0.66666667 0.9709506 0.64730040
#> 6   0.11111111 0.4689956 0.05211062
#> 7   0.00000000 0.0000000 0.00000000
#> 8   0.00000000 0.0000000 0.00000000
#> 9   0.00000000 0.0000000 0.00000000
#> 10  0.00000000 0.0000000 0.00000000
#> 11  0.00000000 0.0000000 0.00000000
#> 12  0.00000000 0.0000000 0.00000000
#> 13  0.00000000 0.0000000 0.00000000
#> 14  0.00000000 0.0000000 0.00000000
#> 15  0.00000000 0.0000000 0.00000000
#> 16  0.00000000 0.0000000 0.00000000
#> 17  0.00000000 0.0000000 0.00000000
#> 18  0.13333333 0.5435644 0.07247526
#> 19  0.28571429 0.8366407 0.23904021
#> 20  0.53846154 1.0000000 0.53846154
#> 21  0.07692308 0.3712323 0.02855633
#> 22  0.00000000 0.0000000 0.00000000
#> 23  0.00000000 0.0000000 0.00000000
#> 24  0.00000000 0.0000000 0.00000000
#> 25  0.00000000 0.0000000 0.00000000
#> 26  0.00000000 0.0000000 0.00000000
#> 27  0.00000000 0.0000000 0.00000000
#> 28  0.00000000 0.0000000 0.00000000
#> 29  0.00000000 0.0000000 0.00000000
#> 30  0.00000000 0.0000000 0.00000000
#> 31  0.00000000 0.0000000 0.00000000
#> 32  0.00000000 0.0000000 0.00000000
#> 33  0.00000000 0.0000000 0.00000000
#> 34  0.11764706 0.5032583 0.05920686
#> 35  0.33333333 0.8960382 0.29867941
#> 36  0.40000000 0.9544340 0.38177360
#> 37  0.23529412 0.7642045 0.17981283
#> 38  0.16666667 0.6292492 0.10487487
#> 39  0.17647059 0.6500224 0.11470984
#> 40  0.13333333 0.5435644 0.07247526
#> 41  0.07692308 0.3712323 0.02855633
#> 42  0.00000000 0.0000000 0.00000000
#> 43  0.00000000 0.0000000 0.00000000
#> 44  0.00000000 0.0000000 0.00000000
#> 45  0.00000000 0.0000000 0.00000000
#> 46  0.00000000 0.0000000 0.00000000
#> 47  0.00000000 0.0000000 0.00000000
#> 48  0.00000000 0.0000000 0.00000000
#> 49  0.11764706 0.5032583 0.05920686
#> 50  0.26666667 0.8112781 0.21634083
#> 51  0.53333333 1.0000000 0.53333333
#> 52  0.41176471 0.9640788 0.39697361
#> 53  0.38888889 0.9494520 0.36923134
#> 54  0.38888889 0.9494520 0.36923134
#> 55  0.33333333 0.8960382 0.29867941
#> 56  0.14285714 0.5665095 0.08092993
#> 57  0.06666667 0.3372901 0.02248600
#> 58  0.00000000 0.0000000 0.00000000
#> 59  0.00000000 0.0000000 0.00000000
#> 60  0.00000000 0.0000000 0.00000000
#> 61  0.00000000 0.0000000 0.00000000
#> 62  0.00000000 0.0000000 0.00000000
#> 63  0.00000000 0.0000000 0.00000000
#> 64  0.00000000 0.0000000 0.00000000
#> 65  0.06666667 0.3372901 0.02248600
#> 66  0.26666667 0.8112781 0.21634083
#> 67  0.35294118 0.9182958 0.32410441
#> 68  0.38888889 0.9494520 0.36923134
#> 69  0.41176471 0.9640788 0.39697361
#> 70  0.46666667 0.9886994 0.46139306
#> 71  0.69230769 0.9402860 0.65096720
#> 72  0.13333333 0.5435644 0.07247526
#> 73  0.06250000 0.3227570 0.02017231
#> 74  0.00000000 0.0000000 0.00000000
#> 75  0.00000000 0.0000000 0.00000000
#> 76  0.00000000 0.0000000 0.00000000
#> 77  0.00000000 0.0000000 0.00000000
#> 78  0.00000000 0.0000000 0.00000000
#> 79  0.00000000 0.0000000 0.00000000
#> 80  0.00000000 0.0000000 0.00000000
#> 81  0.13333333 0.5435644 0.07247526
#> 82  0.17647059 0.6500224 0.11470984
#> 83  0.16666667 0.6292492 0.10487487
#> 84  0.22222222 0.7424876 0.16499724
#> 85  0.33333333 0.8960382 0.29867941
#> 86  0.42857143 0.9709506 0.41612168
#> 87  0.13333333 0.5435644 0.07247526
#> 88  0.00000000 0.0000000 0.00000000
#> 89  0.00000000 0.0000000 0.00000000
#> 90  0.00000000 0.0000000 0.00000000
#> 91  0.00000000 0.0000000 0.00000000
#> 92  0.00000000 0.0000000 0.00000000
#> 93  0.00000000 0.0000000 0.00000000
#> 94  0.00000000 0.0000000 0.00000000
#> 95  0.00000000 0.0000000 0.00000000
#> 96  0.00000000 0.0000000 0.00000000
#> 97  0.00000000 0.0000000 0.00000000
#> 98  0.05555556 0.2974722 0.01652624
#> 99  0.11764706 0.5032583 0.05920686
#> 100 0.26666667 0.8112781 0.21634083
#> 101 0.53846154 1.0000000 0.53846154
#> 102 0.20000000 0.6962123 0.13924245
#> 103 0.06250000 0.3227570 0.02017231
#> 104 0.00000000 0.0000000 0.00000000
#> 105 0.00000000 0.0000000 0.00000000
#> 106 0.00000000 0.0000000 0.00000000
#> 107 0.00000000 0.0000000 0.00000000
#> 108 0.00000000 0.0000000 0.00000000
#> 109 0.00000000 0.0000000 0.00000000
#> 110 0.00000000 0.0000000 0.00000000
#> 111 0.00000000 0.0000000 0.00000000
#> 112 0.05882353 0.3095434 0.01820844
#> 113 0.16666667 0.6292492 0.10487487
#> 114 0.33333333 0.8997438 0.29991459
#> 115 0.53333333 1.0000000 0.53333333
#> 116 0.28571429 0.8366407 0.23904021
#> 117 0.06666667 0.3372901 0.02248600
#> 118 0.00000000 0.0000000 0.00000000
#> 119 0.00000000 0.0000000 0.00000000
#> 120 0.00000000 0.0000000 0.00000000
#> 121 0.00000000 0.0000000 0.00000000
#> 122 0.00000000 0.0000000 0.00000000
#> 123 0.00000000 0.0000000 0.00000000
#> 124 0.00000000 0.0000000 0.00000000
#> 125 0.00000000 0.0000000 0.00000000
#> 126 0.00000000 0.0000000 0.00000000
#> 127 0.05882353 0.3095434 0.01820844
#> 128 0.16666667 0.6292492 0.10487487
#> 129 0.76470588 0.8524052 0.65183925
#> 130 0.53333333 1.0000000 0.53333333
#> 131 0.23076923 0.7495953 0.17298352
#> 132 0.06666667 0.3372901 0.02248600
#> 133 0.00000000 0.0000000 0.00000000
#> 134 0.00000000 0.0000000 0.00000000
#> 135 0.00000000 0.0000000 0.00000000
#> 136 0.00000000 0.0000000 0.00000000
#> 137 0.00000000 0.0000000 0.00000000
#> 138 0.00000000 0.0000000 0.00000000
#> 139 0.00000000 0.0000000 0.00000000
#> 140 0.00000000 0.0000000 0.00000000
#> 141 0.00000000 0.0000000 0.00000000
#> 142 0.11764706 0.5032583 0.05920686
#> 143 0.27777778 0.8314744 0.23096511
#> 144 0.55555556 0.9980009 0.55444494
#> 145 0.33333333 0.8960382 0.29867941
#> 146 0.00000000 0.0000000 0.00000000
#> 147 0.00000000 0.0000000 0.00000000
#> 148 0.00000000 0.0000000 0.00000000
#> 149 0.00000000 0.0000000 0.00000000
#> 150 0.00000000 0.0000000 0.00000000
#> 151 0.00000000 0.0000000 0.00000000
#> 152 0.00000000 0.0000000 0.00000000
#> 153 0.00000000 0.0000000 0.00000000
#> 154 0.00000000 0.0000000 0.00000000
#> 155 0.00000000 0.0000000 0.00000000
#> 156 0.00000000 0.0000000 0.00000000
#> 157 0.00000000 0.0000000 0.00000000
#> 158 0.16666667 0.6292492 0.10487487
#> 159 0.35294118 0.9182958 0.32410441
#> 160 0.46666667 0.9886994 0.46139306
#> 161 0.07692308 0.3712323 0.02855633
#> 162 0.00000000 0.0000000 0.00000000
#> 163 0.00000000 0.0000000 0.00000000
#> 164 0.00000000 0.0000000 0.00000000
#> 165 0.00000000 0.0000000 0.00000000
#> 166 0.00000000 0.0000000 0.00000000
#> 167 0.00000000 0.0000000 0.00000000
#> 168 0.00000000 0.0000000 0.00000000
#> 169 0.00000000 0.0000000 0.00000000
#> 170 0.00000000 0.0000000 0.00000000
#> 171 0.00000000 0.0000000 0.00000000
#> 172 0.00000000 0.0000000 0.00000000
#> 173 0.16666667 0.6292492 0.10487487
#> 174 0.38888889 0.9494520 0.36923134
#> 175 0.46666667 0.9886994 0.46139306
#> 176 0.14285714 0.5665095 0.08092993
#> 177 0.00000000 0.0000000 0.00000000
#> 178 0.00000000 0.0000000 0.00000000
#> 179 0.00000000 0.0000000 0.00000000
#> 180 0.00000000 0.0000000 0.00000000
#> 181 0.00000000 0.0000000 0.00000000
#> 182 0.00000000 0.0000000 0.00000000
#> 183 0.00000000 0.0000000 0.00000000
#> 184 0.00000000 0.0000000 0.00000000
#> 185 0.00000000 0.0000000 0.00000000
#> 186 0.00000000 0.0000000 0.00000000
#> 187 0.00000000 0.0000000 0.00000000
#> 188 0.00000000 0.0000000 0.00000000
#> 189 0.11764706 0.5032583 0.05920686
#> 190 0.33333333 0.8960382 0.29867941
#> 191 0.38461538 0.9402860 0.36164845
#> 192 0.06666667 0.3372901 0.02248600
#> 193 0.00000000 0.0000000 0.00000000
#> 194 0.00000000 0.0000000 0.00000000
#> 195 0.00000000 0.0000000 0.00000000
#> 196 0.00000000 0.0000000 0.00000000
#> 197 0.00000000 0.0000000 0.00000000
#> 198 0.00000000 0.0000000 0.00000000
#> 199 0.00000000 0.0000000 0.00000000
#> 200 0.00000000 0.0000000 0.00000000
#> 201 0.00000000 0.0000000 0.00000000
#> 202 0.00000000 0.0000000 0.00000000
#> 203 0.00000000 0.0000000 0.00000000
#> 204 0.16666667 0.6292492 0.10487487
#> 205 0.33333333 0.8960382 0.29867941
#> 206 0.42857143 0.9709506 0.41612168
#> 207 0.06666667 0.3372901 0.02248600
#> 208 0.00000000 0.0000000 0.00000000
#> 209 0.00000000 0.0000000 0.00000000
#> 210 0.00000000 0.0000000 0.00000000
#> 211 0.00000000 0.0000000 0.00000000
#> 212 0.00000000 0.0000000 0.00000000
#> 213 0.00000000 0.0000000 0.00000000
#> 214 0.00000000 0.0000000 0.00000000
#> 215 0.00000000 0.0000000 0.00000000
#> 216 0.00000000 0.0000000 0.00000000
#> 217 0.00000000 0.0000000 0.00000000
#> 218 0.00000000 0.0000000 0.00000000
#> 219 0.00000000 0.0000000 0.00000000
#> 220 0.07692308 0.3712323 0.02855633
#> 221 0.50000000 0.9957275 0.49786373
#> 222 0.30769231 0.8631206 0.26557556
#> 223 0.14285714 0.5665095 0.08092993
#> 224 0.00000000 0.0000000 0.00000000
#> 225 0.00000000 0.0000000 0.00000000
#> 226 0.00000000 0.0000000 0.00000000
#> 227 0.00000000 0.0000000 0.00000000
#> 228 0.00000000 0.0000000 0.00000000
#> 229 0.00000000 0.0000000 0.00000000
#> 230 0.00000000 0.0000000 0.00000000
#> 231 0.00000000 0.0000000 0.00000000
#> 232 0.00000000 0.0000000 0.00000000
#> 233 0.00000000 0.0000000 0.00000000
#> 234 0.00000000 0.0000000 0.00000000
#> 235 0.11111111 0.4689956 0.05211062
#> 236 0.55555556 1.0000000 0.55555556
#> 237 0.22222222 0.7219281 0.16042847
#> 238 0.10000000 0.4394970 0.04394970
#> 239 0.00000000 0.0000000 0.00000000
#> 240 0.00000000 0.0000000 0.00000000